Added the python files
The python section of the project is made, for now, by a main.py, requirements.txt and a Dockerfile.

As for the section future, it'll be dockerized as to be easy to build this project in another computer without any problem or constraint and to better get a replicate of the results, since this project is to demonstrate the application of the mathematical/robotics concepts through a programming language and a set of parameters/data.