Starting project recreation
Now the basis for each section of the project may be lay down for development and, therefore, its recreation.

For that, it's essential to have symbolic variables for the different operations and formulations that there is to make in the Kinematics part - rotations and Jacobian matrices - as well in the Dynamics part - the B matrix, phi and g vectors, such as well the tau vector.

Besides that, there is an explanation in the README, with current status. This is for helping in the project formulation and points of situation throughout the project development.

Maybe, the following commits will be made inside a branch (maybe besides the ones with the CI elements), dedicated each one for julia or for python, not for the same project.