DHC Table added in Python version
For starters, the DHC table is made inside a function in the dhc/table.py module, with output as the table itself - as Robot - and the q's symbolic variables in an array (in this case, the array is in form q = (q1, q2, q3, q4, q5, q6, q7)). This table, however, is only made for kinematic porpuses, since afterwards it will be necessary to generate one table to the dynamic's part, which will derive from this one while adding new columns to the equation.

The goal for now is to output the Direct Kinematic Rotation Matrix and the end-effector position. Afterwards, it'll be necessary to verify how to get a behavior similar to the SIMULINK one. Maybe testing with simupy or get a different approach to the problem, where the block representation isn't going to be necessary, but another representation may be more efficient.