DHC Table added in Julia version
As in MATLAB, now there is a function which creates a matrix with the DHC parameters for the project's robot. In this file, the Robot and the q vector is outputed, which the latter be in the form of q = [q1(t), q2(t), q3(t), q4(t), q5(t), q6(t), q7(t)] for easy access in for-loops or while-loops.

For now, the matrix is done and the next step is to create the Direct Kinematics Rotation matrix and the end-effector position in relation the base frame. After that, it will be needed to create a way to simulate the process just like SIMULINK or develop a new way of simulating, maybe using the REPL or to generate an UI that can change the q's values and see if the mathematical procedure is done correctly as well done as in the simulink version.