Python's Kinematic tests added
This implementation enables in the project the different components validation, proving it that the scripts and modules developed up until some time works and outputs the expected outputs. Besides that, this checks that the future improvements in the project don't broke the already written code, making this a future proof addition.