Added the test for the question 2 in kinematics of the Python version
The DHC table was modified due of representing errors in its formulation and it was added two more tests as requested in the project formulation. These tests checks if the end-effector position and orientation corresponds to a known position - where in this case, it's the arm positioned in the vertical configuration, being the pe only in the z-axis (pe=[0,0,1.2]) and its orientation is just a 90º degree angle rotation in the z-axis of the base frame (R=[0,-1,0;1,0,0;0,0,1]).